A rough start of a Check-In Maintenance screen.
The URL is /barcodes.

This will be used by Cedar Ridge to set family barcode numbers for
our new check-in system. The maintenance and setup interface will
be in OneBody, while the actual check-in app itself will be separate.

This could be useful to other churches, but won't be billed as a
"feature" since it is highly specific to our church at the moment.
If you're interested in building on this for your own church,
please let me know (tim@timmorgan.org) so we can possibly collaborate.