Use symbol for form_for to speed up these partials.
This is interesting. For a reason I don't understand right now, instantiating a new
model instance (even if you don't save it to the db) slows down the form_for helper
versus using a symbol and specifying the url manually. I assume something is happening
with the database, causing the slowdown.