Remove Facebook feed import icons.
It's still possible to import from a Facebook status feed, but it now
only works for "old" Facebook users. Users with an ID higher than a
certain unknown number (seems to be users created within the last year
or so) do not have an operational status feed, so this feature does not
work for them.

Sorry. :-(