Native sync with PowerChurch (one-way).
In PowerChurch, in the "OneBody Sync Settings" window, under the "Server" tab,
paste your api key into the "OneBody API key" field (you must be an admin user
with the import_data privilege); for the "Sync URL" field, enter the URL
for your OneBody site, then a slash, then "pc_sync" -- like so:

    http://example.com/pc_sync

The sync process is offloaded as much as possible to the background. Once
PowerChurch has finished uploading zip files, OneBody spawns a runner
process that unzips the files and syncs the data.

You can see the outcome of a sync job by going to the Admin dashboard
and scrolling down to the "Sync" heading. Also, if the last sync job
resulted in any errors, you will see an alert at the top of the Admin
dashboard.